cf_apiary/requirements.txt
google-cloud-logging==1.11.0
google-api-python-client==1.7.9
google-cloud-trace==0.21.0
opencensus==0.6.0
opencensus-context==0.1.1
opencensus-ext-stackdriver==0.4.0
